unk8sify a bunch of pps cmd test
