+++ src/server/pps/cmds/BUILD.bazel
@@ -50,16 +50,22 @@ go_test(
     ],
     embed = [":cmds"],
     pure = "on",
+    shard_count = 4,
     deps = [
         "//src/admin",
+        "//src/internal/client",
         "//src/internal/dockertestenv",
         "//src/internal/errors",
+        "//src/internal/pachd",
         "//src/internal/pctx",
         "//src/internal/ppsutil",
         "//src/internal/require",
         "//src/internal/testpachd/realenv",
         "//src/internal/testutil",
         "//src/pps",
+        "@io_k8s_api//core/v1:core",
+        "@io_k8s_apimachinery//pkg/apis/meta/v1:meta",
+        "@io_k8s_client_go//kubernetes",
         "@org_golang_google_protobuf//encoding/protojson",
     ],
 )
